# src/main/java/org/superquinquin/releve/ReleveDto.java
package org.superquinquin.releve;

import org.eclipse.microprofile.openapi.annotations.media.Schema;

import java.time.LocalDate;
import java.util.List;

@Schema(requiredProperties = {"date", "lines"})
public record ReleveDto(LocalDate date, List<ReleveLineDto> lines) {
}